+++ src/cnh/index.ts
@@ -11,7 +11,7 @@ import { clean } from '../utils/clean.js';
  * @param value O número da CNH a ser validado, com ou sem máscara.
  * @returns `true` se o formato da CNH for válido, `false` caso contrário.
  */
-export function cnh(value: string | number): boolean {
+export function isValidCnh(value: string | number): boolean {
   const cleanedValue = clean(value);
 
   if (cleanedValue.length !== 11 || /^(\d)\1{10}$/.test(cleanedValue)) {